requirements/dev.txt
altgraph==0.17.4
cachetools==5.5.2
certifi==2025.1.31
cfgv==3.4.0
charset-normalizer==3.4.1
distlib==0.3.9
filelock==3.16.1
google-api-core==2.24.2
google-api-python-client==2.167.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.70.0
httplib2==0.22.0
identify==2.6.1
idna==3.10
nodeenv==1.9.1
oauthlib==3.2.2
packaging==24.2
platformdirs==4.3.6
pre-commit==3.5.0
proto-plus==1.26.1
protobuf==6.30.2
pyasn1==0.6.1
pyasn1-modules==0.4.2
pyinstaller==6.13.0
pyinstaller-hooks-contrib==2025.3
pyparsing==3.2.3
pyside6==6.9.0
pyside6-addons==6.9.0
pyside6-essentials==6.9.0
pyyaml==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
shiboken6==6.9.0
uritemplate==4.1.1
urllib3==2.4.0
virtualenv==20.30.0

# The following packages are considered to be unsafe in a requirements file:
# setuptools
